+++ semana19/projeto/src/business/services/authenticator.ts
@@ -1,0 +1,26 @@
+import * as jwt from "jsonwebtoken"
+import { AuthenticationData } from "../entities/types"
+
+
+export function generateToken(
+    payload: AuthenticationData
+): string {
+    return jwt.sign(
+        payload,
+        process.env.JWT_KEY as string,
+        {
+            expiresIn: process.env.JWT_EXPIRE_TIME
+        }
+    )
+}
+
+export function getTokenData(
+    token: string
+): AuthenticationData {
+    const result: any = jwt.verify(
+        token,
+        process.env.JWT_KEY as string
+    )
+
+    return { id: result.id, }
+}
