+++ semana19/projeto/src/controller/routes/postRoutes.ts
@@ -1,0 +1,8 @@
+import express from "express";
+import { createPost, getPostById } from "../postController";
+
+
+export const postRouter = express.Router()
+
+postRouter.post("/create", createPost)
+postRouter.get("/:id", getPostById)
