+++ semana19/projeto/src/controller/routes/userRoutes.ts
@@ -1,0 +1,7 @@
+import express from "express";
+import { login, signup } from "../userController";
+
+export const userRouter = express.Router()
+
+userRouter.post("/signup", signup)
+userRouter.post("/login", login)
