+++ semana19/projeto/src/data/configurations/connection.ts
@@ -1,0 +1,16 @@
+import knex from "knex"
+import dotenv from "dotenv"
+import Knex from "knex"
+
+dotenv.config()
+
+export const connection: Knex = knex({
+   client: "mysql",
+   connection: {
+      host: process.env.DB_HOST,
+      user: process.env.DB_USER,
+      password: process.env.DB_PASSWORD,
+      port: 3306,
+      database: process.env.DB_NAME,
+   }
+})
